requirements.txt
pytesseract 
pillow